Move DELF resizing code into utils.py, which is more appropriate than extractor.py. (#9108)
* Merged commit includes the following changes:
326723075  by Andre Araujo:

    Move image resize utility into utils.py.

--

PiperOrigin-RevId: 326723075

* Adding back matched_images_demo.png
